Altered createGraph method to read from text file
proposed a textfile labeled "CLASS_INFO" to be added to the repository. this text file holds the course info as written in the project file.

this change to the project should allow the graph to be generated by reading the "CLASS_INFO" file instead of hardcoding it into the project. I am not certain if the path to the file is correct though. But, based on my personal texting it should run with no errors if the path is correct though.